feat: canvas store
